Fix names, methods, return values to match API
This constantly throws me off - it's easier if classes and methods have
the same names as the S3 API the implement.
