+++ server/src/main/java/com/adobe/testing/s3mock/dto/Bucket.java
@@ -18,11 +18,13 @@ package com.adobe.testing.s3mock.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonRootName;
 import java.nio.file.Path;
 
 /**
  * <a href="https://docs.aws.amazon.com/AmazonS3/latest/API/API_Bucket.html">API Reference</a>.
  */
+@JsonRootName("Bucket")
 public class Bucket {
 
   @JsonProperty("Name")
