+++ server/src/main/java/com/adobe/testing/s3mock/dto/CompletedPart.java
@@ -17,12 +17,14 @@
 package com.adobe.testing.s3mock.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonRootName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 /**
  * <a href="https://docs.aws.amazon.com/AmazonS3/latest/API/API_CompletedPart.html">API Reference</a>.
  */
+@JsonRootName("CompletedPart")
 public class CompletedPart {
 
   @JsonProperty("PartNumber")
