+++ server/src/test/resources/com/adobe/testing/s3mock/dto/DeleteResultTest_testSerialization.xml
