+++ server/src/test/resources/com/adobe/testing/s3mock/dto/DeleteTest_testDeserialization.xml
